+++ Plugins/Rhetos.LightDMS/Storage/S3Storage.cs
@@ -1,17 +1,25 @@
 ﻿using Amazon.S3;
 using Amazon.S3.Transfer;
+using Rhetos.LightDMS;
 
 namespace Rhetos.LightDms.Storage
 {
-    public static class S3StorageClient
+    public class S3StorageClient
     {
-        public static AmazonS3Client GetAmazonS3Client()
+        private readonly S3Options _s3Options;
+
+        public S3StorageClient(S3Options s3Options)
         {
-            var key = System.Configuration.ConfigurationManager.AppSettings.Get("LightDms.S3.Key");
-            var accessKeyID = System.Configuration.ConfigurationManager.AppSettings.Get("LightDms.S3.AccessKeyID");
+            _s3Options = s3Options;
+        }
+
+        public AmazonS3Client GetAmazonS3Client()
+        {
+            var key = _s3Options.Key;
+            var accessKeyID = _s3Options.AccessKeyID;
             var s3Config = new AmazonS3Config()
             {
-                ServiceURL = System.Configuration.ConfigurationManager.AppSettings.Get("LightDms.S3.ServiceURL")
+                ServiceURL = _s3Options.ServiceURL
             };
             if (string.IsNullOrWhiteSpace(key) || string.IsNullOrWhiteSpace(accessKeyID) || string.IsNullOrWhiteSpace(s3Config.ServiceURL))
                 throw new FrameworkException("Invalid S3 storage configuration parameters.");